pr comments

+++ api/v1alpha1/clienttrafficpolicy_types.go
@@ -88,11 +88,13 @@ type ClientTrafficPolicySpec struct {
 	//
 	// +optional
 	Timeout *ClientTimeout `json:"timeout,omitempty"`
-	// PerConnectionBufferLimitBytes sets a soft limit on size of the cluster’s connections read and write buffers.
+	// PerConnectionBufferLimitBytes sets a soft limit on size of the listener’s new connection read and write buffers.
 	// Default: 32768.
 	//
+	// +kubebuilder:validation:Minimum=0
+	// +kubebuilder:validation:Maximum=4294967295
 	// +optional
-	PerConnectionBufferLimitBytes *uint32 `json:"perConnectionBufferLimitBytes,omitempty"`
+	PerConnectionBufferLimitBytes *int64 `json:"perConnectionBufferLimitBytes,omitempty"`
 }
 
 // HeaderSettings providess configuration options for headers on the listener.
